+++ blaze-core/src/test/scala/org/http4s/blazecore/util/Http1WriterSpec.scala
@@ -223,13 +223,13 @@ class Http1WriterSpec extends Http4sSpec {
     }
 
     val resource: Stream[IO, Byte] =
-      bracket(IO("foo"))({ str =>
+      bracket(IO("foo"))(_ => IO.unit).flatMap { str =>
         val it = str.iterator
         emit {
           if (it.hasNext) Some(it.next.toByte)
           else None
         }
-      }, _ => IO.unit).unNoneTerminate
+      }.unNoneTerminate
 
     "write a resource" in {
       val p = resource