+++ jawn/src/main/scala/org/http4s/jawn/JawnInstances.scala
@@ -1,7 +1,7 @@
 package org.http4s
 package jawn
 
-import _root_.jawn.{AsyncParser, Facade, ParseException}
+import _root_.jawn.{AsyncParser, ParseException, RawFacade}
 import cats.effect._
 import cats.syntax.either._
 import cats.syntax.functor._
@@ -9,11 +9,12 @@ import fs2.Stream
 import jawnfs2._
 
 trait JawnInstances {
-  def jawnDecoder[F[_]: Sync, J: Facade]: EntityDecoder[F, J] =
+  def jawnDecoder[F[_]: Sync, J: RawFacade]: EntityDecoder[F, J] =
     EntityDecoder.decodeBy(MediaType.application.json)(jawnDecoderImpl[F, J])
 
   // some decoders may reuse it and avoid extra content negotiation
-  private[http4s] def jawnDecoderImpl[F[_]: Sync, J: Facade](msg: Message[F]): DecodeResult[F, J] =
+  private[http4s] def jawnDecoderImpl[F[_]: Sync, J: RawFacade](
+      msg: Message[F]): DecodeResult[F, J] =
     DecodeResult {
       msg.body.chunks
         .parseJson(AsyncParser.SingleValue)