+++ server/src/main/scala/org/http4s/server/ServerBuilder.scala
@@ -3,13 +3,13 @@ package server
 
 import cats.implicits._
 import cats.effect._
+import cats.effect.concurrent.Ref
 import fs2.StreamApp.ExitCode
 import fs2._
 import java.net.{InetAddress, InetSocketAddress}
 import java.util.concurrent.ExecutorService
 import javax.net.ssl.SSLContext
 import fs2.async.immutable.Signal
-import fs2.async.Ref
 import org.http4s.server.SSLKeyStoreSupport.StoreInfo
 import scala.collection.immutable
 import scala.concurrent.ExecutionContext
@@ -56,10 +56,10 @@ trait ServerBuilder[F[_]] {
     * Runs the server as a process that never emits.  Useful for a server
     * that runs for the rest of the JVM's life.
     */
-  final def serve(implicit F: Effect[F], ec: ExecutionContext): Stream[F, ExitCode] =
+  final def serve(implicit F: Concurrent[F]): Stream[F, ExitCode] =
     for {
       signal <- Stream.eval(async.signalOf[F, Boolean](false))
-      exitCode <- Stream.eval(async.refOf[F, ExitCode](ExitCode.Success))
+      exitCode <- Stream.eval(Ref[F].of(ExitCode.Success))
       serve <- serveWhile(signal, exitCode)
     } yield serve
 
@@ -70,11 +70,9 @@ trait ServerBuilder[F[_]] {
   final def serveWhile(
       terminateWhenTrue: Signal[F, Boolean],
       exitWith: Ref[F, ExitCode]): Stream[F, ExitCode] =
-    Stream.bracket(start)(
-      (_: Server[F]) =>
-        terminateWhenTrue.discrete.takeWhile(_ === false).drain ++ Stream.eval(exitWith.get),
-      _.shutdown
-    )
+    Stream.bracket(start)(_.shutdown) *> (terminateWhenTrue.discrete
+      .takeWhile(_ === false)
+      .drain ++ Stream.eval(exitWith.get))
 
   /** Set the banner to display when the server starts up */
   def withBanner(banner: immutable.Seq[String]): Self