+++ server/src/main/scala/org/http4s/server/middleware/RequestLogger.scala
@@ -4,11 +4,11 @@ package middleware
 
 import cats.data.{Kleisli, OptionT}
 import cats.effect._
+import cats.effect.concurrent.Ref
 import cats.implicits._
 import fs2._
 import org.http4s.util.CaseInsensitiveString
 import org.log4s._
-import scala.concurrent.ExecutionContext
 
 /**
   * Simple Middleware for Logging Requests As They Are Processed
@@ -16,19 +16,18 @@ import scala.concurrent.ExecutionContext
 object RequestLogger {
   private[this] val logger = getLogger
 
-  def apply[F[_]: Effect](
+  def apply[F[_]: Concurrent](
       logHeaders: Boolean,
       logBody: Boolean,
       redactHeadersWhen: CaseInsensitiveString => Boolean = Headers.SensitiveHeaders.contains
-  )(@deprecatedName('service) routes: HttpRoutes[F])(
-      implicit ec: ExecutionContext = ExecutionContext.global): HttpRoutes[F] =
+  )(@deprecatedName('service) routes: HttpRoutes[F]): HttpRoutes[F] =
     Kleisli { req =>
       if (!logBody)
         OptionT(Logger
           .logMessage[F, Request[F]](req)(logHeaders, logBody)(logger.info(_)) *> routes(req).value)
       else
         OptionT
-          .liftF(async.refOf[F, Vector[Segment[Byte, Unit]]](Vector.empty[Segment[Byte, Unit]]))
+          .liftF(Ref[F].of(Vector.empty[Segment[Byte, Unit]]))
           .flatMap { vec =>
             val newBody = Stream
               .eval(vec.get)
@@ -38,7 +37,7 @@ object RequestLogger {
             val changedRequest = req.withBodyStream(
               req.body
               // Cannot Be Done Asynchronously - Otherwise All Chunks May Not Be Appended Previous to Finalization
-                .observe(_.segments.flatMap(s => Stream.eval_(vec.modify(_ :+ s))))
+                .observe(_.segments.flatMap(s => Stream.eval_(vec.update(_ :+ s))))
             )
 
             val response = routes(changedRequest)