+++ server/src/main/scala/org/http4s/server/middleware/ResponseLogger.scala
@@ -4,11 +4,11 @@ package middleware
 
 import cats.data.{Kleisli, OptionT}
 import cats.effect._
+import cats.effect.concurrent.Ref
 import cats.implicits._
 import fs2._
 import org.http4s.util.CaseInsensitiveString
 import org.log4s.getLogger
-import scala.concurrent.ExecutionContext
 
 /**
   * Simple middleware for logging responses as they are processed
@@ -20,9 +20,7 @@ object ResponseLogger {
       logHeaders: Boolean,
       logBody: Boolean,
       redactHeadersWhen: CaseInsensitiveString => Boolean = Headers.SensitiveHeaders.contains
-  )(@deprecatedName('service) routes: HttpRoutes[F])(
-      implicit F: Effect[F],
-      ec: ExecutionContext = ExecutionContext.global): HttpRoutes[F] =
+  )(@deprecatedName('service) routes: HttpRoutes[F])(implicit F: Concurrent[F]): HttpRoutes[F] =
     Kleisli { req =>
       routes(req)
         .semiflatMap { response =>
@@ -30,24 +28,23 @@ object ResponseLogger {
             Logger.logMessage[F, Response[F]](response)(logHeaders, logBody, redactHeadersWhen)(
               logger.info(_)) *> F.delay(response)
           else
-            async.refOf[F, Vector[Segment[Byte, Unit]]](Vector.empty[Segment[Byte, Unit]]).map {
-              vec =>
-                val newBody = Stream
-                  .eval(vec.get)
-                  .flatMap(v => Stream.emits(v).covary[F])
-                  .flatMap(c => Stream.segment(c).covary[F])
+            Ref[F].of(Vector.empty[Segment[Byte, Unit]]).map { vec =>
+              val newBody = Stream
+                .eval(vec.get)
+                .flatMap(v => Stream.emits(v).covary[F])
+                .flatMap(c => Stream.segment(c).covary[F])
 
-                response.copy(
-                  body = response.body
-                  // Cannot Be Done Asynchronously - Otherwise All Chunks May Not Be Appended Previous to Finalization
-                    .observe(_.segments.flatMap(s => Stream.eval_(vec.modify(_ :+ s))))
-                    .onFinalize {
-                      Logger.logMessage[F, Response[F]](response.withBodyStream(newBody))(
-                        logHeaders,
-                        logBody,
-                        redactHeadersWhen)(logger.info(_))
-                    }
-                )
+              response.copy(
+                body = response.body
+                // Cannot Be Done Asynchronously - Otherwise All Chunks May Not Be Appended Previous to Finalization
+                  .observe(_.segments.flatMap(s => Stream.eval_(vec.update(_ :+ s))))
+                  .onFinalize {
+                    Logger.logMessage[F, Response[F]](response.withBodyStream(newBody))(
+                      logHeaders,
+                      logBody,
+                      redactHeadersWhen)(logger.info(_))
+                  }
+              )
             }
         }
         .handleErrorWith(t =>